Patched some errors

+++ Headers/Utils/Constants.h
@@ -6,9 +6,9 @@
 class Constants {
 	
 	public:
-		unsigned short const static MAX_ITER = 10000;
-		unsigned short const static WIDTH = 1920;
-		unsigned short const static HEIGHT = 1080;
+		unsigned int const static MAX_ITER = 10000;
+		unsigned int const static WIDTH = 1920;
+		unsigned int const static HEIGHT = 1080;
 
 		unsigned short const static HALF_HEIGHT = HEIGHT / 2;
 		unsigned short const static HALF_WIDTH = WIDTH / 2;
